requirements.txt
streamlit
transformers
torch
langchain
sentence-transformers
faiss-cpu
pandas
numpy
accelerate
bitsandbytes
